iMostly assigns outputs to pins
This file maps the outputs from the top level module to pins/pads
on the chip. Those pads apear on 4 40 pin headers and are connected
via ribbon cable to the sixteen segment led displays. These are
common annode displays with a couple diodes on the commons to
knock the voltage down a bit.